rpc: rename `lib/` to `jsonrpc/`
