+++ rpc/jsonrpc/client/http_json_client.go
@@ -13,7 +13,7 @@ import (
 
 	amino "github.com/tendermint/go-amino"
 
-	types "github.com/tendermint/tendermint/rpc/lib/types"
+	types "github.com/tendermint/tendermint/rpc/jsonrpc/types"
 )
 
 const (
