+++ rpc/jsonrpc/server/http_server.go
@@ -16,7 +16,7 @@ import (
 	"golang.org/x/net/netutil"
 
 	"github.com/tendermint/tendermint/libs/log"
-	types "github.com/tendermint/tendermint/rpc/lib/types"
+	types "github.com/tendermint/tendermint/rpc/jsonrpc/types"
 )
 
 // Config is a RPC server configuration.
